Phase K: migrate MasterJoin + VersionRequest to int32_t
Both structs are cast directly from the wire buffer on the Linux
dispatch path. With `long` fields they were 108 bytes (MasterJoin) and
16 bytes (VersionRequest) on Linux x86_64 vs 64 / 8 on Win32 — the
client's wire format.

VersionRequest was working by coincidence: the test sends an 8-byte
payload, the Linux 16-byte struct read past it, but `(uint32_t) Major`
truncated to the bottom 4 bytes (matching the wire on LE x86_64) and
Minor read 0 from the zero-padded recv buffer. Making it int32_t
removes the accidental correctness.

MasterJoin is the real fix. Before this commit, HandleMasterJoin read
avatar_id_lsb at struct offset 32 and ToSectorID at offset 40 (8-byte
field layout), instead of 16 / 20 where they actually sit on the wire.
The handler saw zeros everywhere and fell through to the
SendMasterLogin-timeout path on every join — proxy log read
"avatar_id=0 ToSectorID=0" no matter what the client sent.

After: replay test (real captured Win32 client bytes) logs
"avatar_id=4150549696 ToSectorID=45151" — the recorded values; and the
synthesized master_join_test logs "avatar_id=1 ToSectorID=1" matching
its stamp pattern.

8/8 integration tests still green.

+++ proxy/PacketStructures.h
@@ -13,8 +13,8 @@ struct EnbTcpHeader
 
 struct VersionRequest
 {
-    long    Major;      // 4 bytes
-    long    Minor;      // 4 bytes
+    int32_t Major;      // 4 bytes on every platform (Phase K — was `long`,
+    int32_t Minor;      // 4 bytes   which is 8 bytes on Linux x86_64)
 } ATTRIB_PACKED;
 
 struct EnbUdpHeader
@@ -226,17 +226,23 @@ struct GlobalCreateCharacter
 
 struct MasterJoin
 {
-    long    unknown1;
-    long    unknown2;
-    long    unknown3;
-    long    avatar_id_msb;
-    long    avatar_id_lsb;
-    long    ToSectorID;
-    long    FromSectorID;
-    long    PlayerLevel;
-    long    unknown8;
-    long    unknown9;
-    long    unknown10;
+    // Phase K: 11 * int32_t + 20-byte ticket = 64 bytes on every platform.
+    // Was `long`-based, which is 8 bytes on Linux x86_64 (struct = 108B) vs
+    // 4 bytes on Win32 (struct = 64B). The wire format is 64 bytes; the Linux
+    // mismatch caused HandleMasterJoin to read avatar_id_lsb / ToSectorID from
+    // bytes 32 / 40 instead of 16 / 20, yielding zeros and triggering the
+    // SendMasterLogin-timeout fallback path on every join.
+    int32_t unknown1;
+    int32_t unknown2;
+    int32_t unknown3;
+    int32_t avatar_id_msb;
+    int32_t avatar_id_lsb;
+    int32_t ToSectorID;
+    int32_t FromSectorID;
+    int32_t PlayerLevel;
+    int32_t unknown8;
+    int32_t unknown9;
+    int32_t unknown10;
     char    ticket[20];
 } ATTRIB_PACKED;
